ath11k: blackrock: Add preliminary definition for WDK2023
Windows Dev Kit 2023 aka "Blackrock" aka "Project Volterra"

This additional definition works better than the default one that is found
for the board string. It is more specific with the string "variant=volterra"
which is the string used in the dt. The calibration data assigned are the
ones from the Lenovo Thinkpad X13s. This is not perfect, but it will get you
~300..500Mbit/s which is better than 36 from the default data.

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>